move PreparedStatementBuilder
