+++ nextstep-jdbc/src/main/java/nextstep/jdbc/template/JdbcTemplate.java
@@ -2,7 +2,6 @@ package nextstep.jdbc.template;
 
 import nextstep.jdbc.exception.JdbcTemplateSqlException;
 import nextstep.jdbc.mapper.JdbcMapper;
-import nextstep.jdbc.query.PreparedStatementBuilder;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
@@ -16,19 +15,19 @@ public class JdbcTemplate {
         this.con = con;
     }
 
-    public <T> T executeQuery(PreparedStatementBuilder preparedStatementBuilder, JdbcMapper<T> mapper) {
-        try (PreparedStatement pstmt = preparedStatementBuilder.create(con);
-             ResultSet rs = pstmt.executeQuery()) {
-
-            return mapper.mapped(rs);
+    public void update(PreparedStatementBuilder preparedStatementBuilder) {
+        try (PreparedStatement pstmt = preparedStatementBuilder.create(con)) {
+            pstmt.executeUpdate();
         } catch (SQLException e) {
             throw new JdbcTemplateSqlException(e);
         }
     }
 
-    public void update(PreparedStatementBuilder preparedStatementBuilder) {
-        try (PreparedStatement pstmt = preparedStatementBuilder.create(con)) {
-            pstmt.executeUpdate();
+    public <T> T executeQuery(PreparedStatementBuilder preparedStatementBuilder, JdbcMapper<T> mapper) {
+        try (PreparedStatement pstmt = preparedStatementBuilder.create(con);
+             ResultSet rs = pstmt.executeQuery()) {
+
+            return mapper.mapped(rs);
         } catch (SQLException e) {
             throw new JdbcTemplateSqlException(e);
         }
